requirements.txt
anyio==3.6.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
click==8.1.3
fastapi==0.94.1
h11==0.14.0
idna==3.4
joblib==1.2.0
lightgbm==3.3.5
numpy==1.24.2
pandas==1.5.3
pydantic==1.10.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
threadpoolctl==3.1.0
typing_extensions==4.5.0
uvicorn==0.21.0
